# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AFRIGAZ",
        "description": "Documentation API AFRIGAZ",
        "contact": {
            "email": "winnersthec001.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/productStoreData": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Créer un produit",
                "operationId": "46093795a53280e87ac16d38a34690a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "gaz"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "wholesale_price": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "retail_price": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "créée"
                    },
                    "422": {
                        "description": "Erreur validation"
                    }
                }
            }
        },
        "/api/v1/productUpdate/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Modifier",
                "operationId": "2e73dc411d03629f7f989b1b7978962b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "gaz"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "wholesale_price": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "retail_price": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "mise à jour"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
        "/api/v1/productDelete/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Supprimer",
                "operationId": "4f004b85bf5654ebd6904e0b34326414",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
...
    },
...
}